refactor: GuestbookElement로 이름 변경
