+++ frontend/src/pages/host/guestbook/list/GuestbookListPage.tsx
@@ -1,4 +1,4 @@
-import GuestbookCard from './card/GuestbookCard';
+import GuestbookElement from './element/GuestbookElement';
 import * as S from './GuestbookListPage.styles';
 
 const GuestbookListPage = () => {
@@ -9,10 +9,10 @@ const GuestbookListPage = () => {
         <S.Description>28명 참여</S.Description>
       </S.InfoContainer>
       <S.ListContainer>
-        <GuestbookCard guestName={'밍고'} hasPhoto={true} isNew={true} />
-        <GuestbookCard guestName={'블루블루'} hasPhoto={true} isNew={true} />
-        <GuestbookCard guestName={'퐁퐁쥬'} hasPhoto={true} isNew={true} />
-        <GuestbookCard guestName={'루나'} hasPhoto={true} isNew={true} />
+        <GuestbookElement guestName={'밍고'} hasPhoto={true} isNew={true} />
+        <GuestbookElement guestName={'블루블루'} hasPhoto={true} isNew={true} />
+        <GuestbookElement guestName={'퐁퐁쥬'} hasPhoto={true} isNew={true} />
+        <GuestbookElement guestName={'루나'} hasPhoto={true} isNew={true} />
       </S.ListContainer>
     </S.Wrapper>
   );
