+++ frontend/src/pages/host/guestbook/list/element/GuestbookElement.styles.ts
